Heavy refactor
Combined five_prime_metaplot, three_prime_metaplot, and divergent_pielup_metaplot to one program called metaplot. Also refactored some of the utility function to their own folders.

+++ GC_bioinfo/utils/heatmap_utils/average_matrix.py
@@ -23,7 +23,7 @@ def average_matrix(filename, num_lines_to_average):
             vals = line.split()
 
             for j, val in enumerate(vals):
-                curr_sums[j] += int(val)
+                curr_sums[j] += float(val)
 
             if (i+1) % num_lines_to_average == 0:
                 # Average and print out
